backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
chromadb==0.4.22
openai==1.12.0
google-generativeai==0.3.2
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
numpy==1.26.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
aiofiles==23.2.1
pytest==8.0.1
httpx==0.26.0 